test(history): split snapshot regressions
